Refactored the public/js Jackalope.js, Unicorn.js, and Yeti.js files into one Creatures.js file.  Also updated the playpen.js file.
