+++ public/assets/lib/Pet.js
@@ -72,6 +72,6 @@ class Pet {
 // Don't forget to export
 // module.exports = Pet;
 
-export {
-    Pet,
-};
+// export {
+//     Pet,
+// };
