+++ views/playpen.handlebars
@@ -1,7 +1,15 @@
 <div class="backgroundGraphic"></div>
 
+<section id="playpen">
+    
+</section>
 
 {{!-- served for the environment --}}
-{{!-- canvas for UI geraphic feed, clean-up buttons
+{{!-- canvas for UI geraphic feed, clean-up buttons --}}
 {{!-- separate JS file for that page --}}
 {{!-- --}}
+{{!--  --}}
+
+<script src="/assets/lib/Pet.js"></script>
+<script src="/assets/lib/Jackalope.js"></script>
+<script src="/assets/js/playpen.js"></script>
